usb: implement fido guest key and interrupt poll
This CL completes the implementation of the virtual fido security key by
adding logic for u2f packet handling and delivery at the guest level,
and adding a poll thread to take care of U2F transactions and XHCI
transfer expiration (to emulate an USB interrupt device).

It also addes the seccomp policies to allow spawning the poll thread in
the sandbox environment.

BUG=b:297957754
TEST=presubmit and unit tests
TEST=manually tested yubico key and it works in the guest

Change-Id: Icedb15702d266fe135d74763602a1c8213df39a2
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5316380
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>