Avoid OnceCell:get_or_try_init()
In preparation for migration to std::sync::OnceLock, replace usage of
get_or_try_init(), since the standard library version of that function
is not stabilized yet. Use get_or_init() in its place, and store the
result of the initialization so future attempts do not need to retry.

Change-Id: I7727b6f9319982054d8633eca1eb1bd2b04c079b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5321448
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>