devices: add notifier for hotplug command complete
When a PCI hot plug or hot unplug event is completed by the guest, the
hardware is notified by PCI_EXP_SLTCTL_CCIE. In this CL, we add a
notifier when hotplug is complete for PcieRootPort. This is infeasible
for other variants of HotPlugBus.

BUG=b:325997618
TEST=manual feature tests
TEST=./tools/dev_container ./tools/presubmit

Change-Id: I6ed138a3a6006b6d02fe660c293e298b1cc04701
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5326701
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Ningyuan Wang <ningyuan@google.com>