cros_tracing_types: statically initialize Mutex
The sync::Mutex::new() function is const now, so it can be constructed
statically rather than requiring lazy initialization.

HashSet::new() is not const, so it cannot be used in a static init;
replace it with the equivalent BTreeSet, which has a const new()
function.

Change-Id: I339903c73f9ab3f41fbb517a30f2987d6d05cb3d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5342657
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>