devices/vmwdt: Poll test result instead of waiting once
In these tests we need to wait for the watchdog to run before
checking the behavior. We do not have a signal for that, so instead
of sleeping once, poll multiple times over the course of a longer
period.

This fixes occasional test flakes because the watchdog has not run
after 100ms.

BUG=b:322053405
TEST=Both tests pass locally

Change-Id: I11fad2c46b0587a2f61b0afc8ecf66206fd3aae8
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5350110
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dennis Kempin <denniskempin@google.com>