devices: merge VhostUserHandler into VhostUserVirtioDevice
The relationship between VhostUserHandler and VhostUserVirtioDevice is
(or has become) unclear. For example, the "handler" is responsible for
starting the worker, but the "device' manages the worker's lifetime. The
division of responsibility for virtio_restore feels somewhat random.
Most of the methods are only forwarding calls to BackendClient.

So, this commit inlines VhostUserHandler into VhostUserVirtioDevice
without any meaningful bevahior changes. In some cases the error
handling is more verbose because some VirtioDevice trait methods don't
return Result.

TEST=started ubuntu VM with a vhost-user block device as the rootfs

Bug: 301269927
Change-Id: Ibc031635f2498fd59d4f868032854da2608c8992
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5384324
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>