devices: move vhost-user frontend to virtio/
Make it clear that the vhost-user frontend is a virtio device and get
rid of some of the extreme levels of nesting.

This is mostly a clean rename, with a small flattening of the
handler/sys mod into just sys.

Also fix up linux -> unix to match the underlying vmm_vhost cfg checks.

Change-Id: Ic737122151641d4742b9815ebcd2ef71070ca1fa
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5386028
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>