devices: vhost-user: align sleep/wake implementation with spec
See the diffs in snapshot_extensions.md for an overview of the changes.

Originally, crosvm's virtio and vhost-user frameworks diverged in their
snapshot-restore implementations. Virtio devices returned ownership of
their queues back to VirtioPciDevice, but vhost-user devices kept
ownership in the backend process. Replacing SLEEP with GET_VRING_BASE
has a side effect of bringing the two into alignment. If you look at
this commit in isolation, you might wonder why we ever went with the
other approach. The reason is that there were a large number of misc
bugs and code organization issue in the way that have been fixed since,
for example, it wasn't possible to restart the vhost-user frontend
worker to rewire the IPC for interrupts, instead we rewired them inside
the backend. Another example: the frontend and backend were not
correctly communicating the current vring base to each other, they
always passed zero.

There is no vhost-user feature bit to indicate if the backend supports
the "suspended device state", so, for devices without support, we no
longer can fail when putting the device to sleep. Instead, we only fail
when we try to snapshot or restore.

Bug: 301269927
Change-Id: I5aaac864e36a2dffe078d7c8f8abff29365a02c0
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5394172
Commit-Queue: Frederick Mayle <fmayle@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>