devices: virtio: return Result from device reset functions
Previously, reset() returned a bool that indicated whether the device
was successfully reset or not; this was inconsistently implemented, and
some devices attempted to handle the condition of being reset while
already not running, which should never happen.

Instead, return a Result<()> which will be Ok(()) if reset succeeded and
Err(...) otherwise, and clean up the redundant checks in reset()
implementations for already-stopped state.

Change-Id: Id48f611081395f95db18ddb0162e31bab4cbb562
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5388498
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Frederick Mayle <fmayle@google.com>