devices: vhost_user_frontend: stop worker on reset
When the device is reset, we now stop the worker and reclaim the backend
request handler. Previously, if the device was activated, reset, and
then activated again, the backend request handler would be lost.

The changes to run_backend_request_handler technically aren't necessary
because all the current await points are safe, so we could just drop the
future, but that might change one day.

This change is also useful for for snapshot-restore if we move away from
the custom sleep/wake message types. On wake, we might be given a new
`Interrupt` and will need to switch the worker to it (the current design
avoids this by making some assumptions and performing these steps during
the restore, which can only happen once).

Why do we stop the backend request worker at all? An alternative
approach is to keep it running all the time and only stop the interrupt
handling worker code. I attempted this, but it felt too complicated for
now.

Bug: 301269927
Change-Id: I8db627f231c1936912904cbf81b7b68cd7a4aa7d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5393642
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Frederick Mayle <fmayle@google.com>