arch: serial: move bus.insert() into add_serial_devices()
The code in sys::add_serial_devices() for adding the device to the bus
is the same in both cases; extract it into the platform-independent
part. This also means the sys code does not need to know about the
serial device index to address mapping (SERIAL_ADDR).

BUG=b:331472317
TEST=tools/dev_container tools/presubmit

Change-Id: Iabf132bd98983c9ce40bd22527e033285668f29f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5399619
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>