devices: Add metrics for virtio and RTC wakeup
Add metrics for virtio and RTC wakeup events that includes how long the
guest was asleep for before the wakeup event occurred. Technically guest
userspace can arm the RTC independent of going to sleep, but in practice
it's only armed by the RTC driver when the guest is suspending.

BUG=b:332466813
TEST=Boot ARCVM w/ArcIdleManager+ArcS2Idle, observe histograms

Change-Id: Id6c684334ebc1e80e4fc342302c617531ce2fb6b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5400364
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: David Stevens <stevensd@chromium.org>