devices: virtio-block: choose executor kind consistently
When `DiskOption::async_executor` isn't specified, `BlockAsync` was
falling back to the crosvm-wide default (which can be overriden with a
CLI flag), but the vhost-user backend executor and the disk image file
were falling back to the handle executor.

It is important that the disk image file and `BlockAsync` use the same
executor kind for the overlapped IO backend to work.

BUG=b:333618911

Change-Id: Iceccd6b7042569644ef301648ea7523f916c6ce5
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5451071
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Frederick Mayle <fmayle@google.com>