devices: pvclock: handle stub -> idle transition.
A state machine was added to pvclock's lifecycle in
0c5a47511137babe665592f0640fdc4b99fe7c69.
This state machine was missing a transition from Stub -> Idle. In the
snapshotting use case, the device is sitting in the Stub state when
we call virtio_sleep, and it needs to transition to Idle so that later
when we restore and wake the device, it has a Tube and can actually
start the Main worker. Since the Stub -> Idle transition was missing,
we ended up in in the None state instead.  Then we'd call virtio_wake
from the None state, which would panic, because virito_wake is only
valid from Idle.

This CL adds the missing state transition. Critically, it also
re-enables virtio-pvclock for the mingw64 feature set. It was
unintentionally turned off because the compile time feature wasn't
added to the right feature group in
0c5a47511137babe665592f0640fdc4b99fe7c69.

BUG=b:333916852
TEST=tested with snapshotting.

Change-Id: I2fdc0cff68d73fd5b4e9a4c463a462139f131067
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5469475
Commit-Queue: Noah Gold <nkgold@google.com>
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Takaya Saeki <takayas@chromium.org>