kvm_sys: clean up arch-specific ioctls
A few x86-specific definitions are moved inside `mod x86`.

The cfg checks for `KVM_GET_REGS` are removed, since this ioctl is in
fact available on aarch64 (arm64); possibly the original confusion was
due to the fact that this ioctl is not available on 32-bit arm, but we
don't support the 32-bit arm ioctl interface.

The arch-specific re-exports are moved to immediately follow the
relevant module for consistency (and to ensure `cargo fmt` doesn't
rearrange them in a confusing way).

Outdated bindgen commands in comments are removed; the bindings are now
generated by an automated script.

BUG=None
TEST=tools/dev_container tools/presubmit

Change-Id: I9b0072e7dc90e7b328652dc39d5d82b30ba8d462
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5518125
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>