aarch64: fix PCI host memory region prefetchable field
The high MMIO region should be marked as prefetchable. This fixes a
kernel warning that resulted from the high MMIO region extending beyond
the range of a 32-bit memory region, which seems to be a limitation of
the Linux kernel PCI host support:

  pci-host-generic 10000.pci: host bridge /pci ranges:
  pci-host-generic 10000.pci:      MEM 0x0002000000..0x0003ffffff -> 0x0002000000
  pci-host-generic 10000.pci:      MEM 0x0226a00000..0xffffffffff -> 0x0226a00000]
  pci-host-generic 10000.pci: Memory resource size exceeds max for 32 bits

This is also just a more accurate representation of the available MMIO
regions. In particular, the GPU's 64-bit prefetchable BAR is allocated
from this region, so this makes the Devicetree description match the PCI
BAR properties.

BUG=None
TEST=Start Crostini and ARCVM on trogdor

Change-Id: I08fd74d5ba2c30525be9064fc996871501941c24
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5523021
Reviewed-by: Noah Gold <nkgold@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>