devices: Check power indicator before attn btn press
When hot unplug on a PCI slot, the slot could already be off. In this
case, pressing attention button would trigger a hot plug attempt on the
guest. This CL adds a check on the slot state before the button press,
and skip the interrupt if it is already off.

BUG=b:331529292
TEST=./tools/dev_container ./tools/presubmit

Change-Id: I5c80e138cf2fb58003121f9efcc693ed26574a10
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/5528836
Commit-Queue: Ningyuan Wang <ningyuan@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>